requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
idna==3.7
minio==7.2.7
OpenCC==0.2
pycparser==2.22
pycryptodome==3.20.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
